Fix input types

+++ treesim.r
@@ -3,7 +3,7 @@ library(argparser, quietly=TRUE)
 
 p = arg_parser("Simulate a species tree with a given height and leaf number. Other parameters are taken from Molloy and Warnow 2018")
 p = add_argument(p, "--l", help="number of leaves", type="integer")
-p = add_argument(p, "--h", help="tree height", type="float")
+p = add_argument(p, "--h", help="tree height", type="double")
 argv = parse_args(p)
 
 br = 10**(-7)
